requirements/base.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.10
Django==5.1.3
django-allauth==65.2.0
django-appconf==1.0.6
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-datatables-too==0.1.4
django-durationwidget==1.0.5
django-durationwidget2==1.0.6
django-js-asset==2.2.0
django-multiupload==0.6.1
django-recaptcha==4.0.0
django-redis==5.4.0
django-select2==8.2.1
django-widget-tweaks==1.5.0
kombu==5.4.2
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
querystring-parser==1.2.4
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
